+++ src/components/layout/home-button/index.js
@@ -1,0 +1,3 @@
+import homeButton from './home-button.vue'
+
+export default homeButton