+++ src/components/views/main/LogFile.vue
@@ -1,6 +1,6 @@
 <template>
   <el-card style="min-height: 350px;">
-    <el-form ref="formValidate" :model="formValidate" :rules="ruleValidate" label-width="120px" :inline="true"
+    <el-form ref="formValidate" :model="formValidate" :rules="ruleValidate" label-width="auto" :inline="true"
              size="mini" @submit.native.prevent>
       <el-form-item :label="$t('forms.startDate')" prop="startDate">
         <el-date-picker type="date" :disabled="form_loading" :disabled-date="disabledDate"